added images deletion

+++ controllers/admin.js
@@ -1,6 +1,7 @@
 const { validationResult } = require("express-validator");
 
 const Product = require("../models/product");
+const fileHelper = require("../util/file");
 
 exports.getAddProduct = (req, res, next) => {
     res.render("admin/edit-product", {
@@ -131,6 +132,7 @@ exports.postEditProduct = (req, res, next) => {
             }
 
             if (image) {
+                fileHelper.deleteFile(product.imageUrl);
                 product.imageUrl = image.path;
             }
 
@@ -177,7 +179,16 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.id;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
+
+    Product.findById(prodId)
+        .then((product) => {
+            if (!product) {
+                return next(new Error("Product Not Found"));
+            }
+            
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({ _id: prodId, userId: req.user._id });
+        })
         .then((result) => {
             console.log("Product has been deleted");
             res.redirect("/admin/products");
